Subversion failed this operation so many times with such obscure error messages that I do not know what the last attempt really did.
FYI The last failure was: 'Error: SSL handshake failed: An existing connection was forcibly closed by the remote host.  '

git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/branches/b_ph_Petr_Hluzin@471 41062d70-b26d-460a-9302-f438a8930db8

+++ src/DlgUpdateBox.cpp
@@ -332,7 +332,7 @@ void CDlgUpdateBox::OnOK()
 
   UpdateToRecord();
 
-  m_NewSymbol->ppp=GetDlgItemInt(UPDATE_PPP);
+  m_NewSymbol->ppp= (BYTE) GetDlgItemInt(UPDATE_PPP);
   m_NewSymbol->SetRecords( m_records );
 
   EndDialog( IDOK );
@@ -730,7 +730,7 @@ void CDlgUpdateBox::OnAddName()
 	// Create a new name that is a copy of the current name...
 	CLibraryStoreSymbol r = getCurrentRecord();
 
-	r.NameID = -1;
+	r.NameID = (DWORD) -1;
 	r.name = "New Symbol";
 	m_records.push_back( r );
